requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
filelock==3.16.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.1
idna==3.7
iniconfig==2.0.0
instructor==1.6.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
nodeenv==1.9.1
numpy==2.1.2
openai==1.52.0
openai-swarm==0.1.1
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
swarm @ git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
tenacity==9.0.0
tqdm==4.66.4
typer==0.12.5
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
virtualenv==20.27.0
Werkzeug==3.0.3
yarl==1.15.5
